[54] extract integration tests from the crate
