+++ src/lib.rs
@@ -1,0 +1,6 @@
+pub mod database;
+mod omni_error;
+pub mod routes;
+pub mod setup;
+mod tournament;
+pub mod users;
