+++ src/routes/teapot.rs
@@ -13,33 +13,3 @@ pub fn route() -> Router<AppState> {
 fn im_a_teapot() -> (StatusCode, &'static str) {
     (StatusCode::IM_A_TEAPOT, IM_A_TEAPOT_RESPONSE)
 }
-
-#[cfg(test)]
-mod tests {
-    use std::future::IntoFuture;
-
-    use reqwest::{Client, StatusCode};
-
-    use crate::{setup::get_socket_addr, test};
-
-    #[tokio::test]
-    async fn test_teapot() {
-        // GIVEN
-        let socket_address = get_socket_addr().to_string();
-        let app = test::create_app().await;
-        let listener = test::create_listener().await;
-        let server = axum::serve(listener, app).into_future();
-        tokio::spawn(server);
-
-        // WHEN
-        let client = Client::new();
-        let res = client
-            .get(format!("http://{}/brew-coffee", socket_address))
-            .send()
-            .await
-            .unwrap();
-
-        // THEN
-        assert_eq!(res.status(), StatusCode::IM_A_TEAPOT);
-    }
-}
